Crate ba2

Archives come in various flavors, and the specific variant you’ll need to use depends on which game you’re working with. Learn more by choosing one of tes3, tes4, or fo4.

If you are uncertain of the origins of your archive, then you may use guess_format to find a starting point.

A note on strings

The Creation Engine absolutely does not handle unicode correctly, and even has some nasty, extant bugs which exist related to characters that utilize the extended ascii range. As such, all strings are marked as binary strings, without encoding (see also BStr or BString). If you must re-encode strings, then, generally speaking, they are encoded using the system code page of whatever computer happened to write the archive. That means English copies of the game are encoded using Windows-1252, Russian copies using Windows-1251, etc. However, this is not a guarantee and is the source of much consternation when writing internationalized applications for the Creation Engine games.

Modules

fo4

Fallout 4

prelude

Convenience using statements for traits that are needed to work with the library.

tes3

TES III: Morrowind

tes4

TES IV: Oblivion

Structs

BStr

A wrapper for &[u8] that provides convenient string oriented trait impls.

BString

A wrapper for Vec<u8> that provides convenient string oriented trait impls.

Borrowed

Makes a shallow copy of the input.

Copied

Makes a deep copy of the input.

Enums

CompressionResult

Indicates whether the operation should finish by compressing the data or not.

FileFormat

The file format for a given archive.

Traits

ByteSlice

A trait that extends &[u8] with string oriented methods.

ByteVec

A trait that extends Vec<u8> with string oriented methods.

CompressableFrom

A trait that creates an optionally compressed container using the given value.

Reader

A trait that enables reading from various sources.

ReaderWithOptions

A trait that enables reading from various sources, with configuration options.

Functions

guess_format

Guesses the archive format for a given source.