ba2
Trait CompressableFrom

pub trait CompressableFrom<T>: Sealed {
    // Required methods
    fn from_compressed(value: T, decompressed_len: usize) -> Self;
    fn from_decompressed(value: T) -> Self;
}

A trait that creates an optionally compressed container using the given value.

Required Methods

fn from_compressed(value: T, decompressed_len: usize) -> Self

Makes a compressed instance of Self using the given data.

fn from_decompressed(value: T) -> Self

Makes a decompressed instance of Self using the given data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.